% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_positive_definite.R
\name{check_if_positive_definite}
\alias{check_if_positive_definite}
\title{check_if_positive_definite}
\usage{
check_if_positive_definite(
  correlation_matrix_as_vector,
  is_lower_triangle,
  has_diagonal
)
}
\arguments{
\item{correlation_matrix_as_vector}{The upper or lower triangle of a correlation matrix, e.g., as reported in an article, formatted as a vector. You can use manual spacing (e.g., linebreaks) to make it appear as the matrix does.}

\item{is_lower_triangle}{If TRUE, the triangle is a lower triangle. If FALSE, it is the upper triangle.}

\item{has_diagonal}{Some articles report a diagonal, either a constant (e.g., 1) or in some fields such as I/O psychology the diagonal reports the scale's reliability (correlation with itself, often Cronbach's alpha).}
}
\value{
logical indicating the matrix is positive-definite (TRUE) or non-positive-definite (FALSE)
}
\description{
Check if a correlation matrix is positive definite, ie whether the correlations are consistent with one another. This function takes a vector as input, and therefore is most useful for extracting and examining matrices extracted from articles.
}
\examples{
# example 1: a lower triangle positive-definite matrix without 1s in diagonal (most common way correlation matrices are reported)
mat1 <- c(-0.85,
          -0.85,  0.90,
          -0.78,  0.83,  0.79,
           0.68, -0.70, -0.71, -0.45,
          -0.87,  0.78,  0.89,  0.66, -0.71)

check_if_positive_definite(mat1, is_lower_triangle = TRUE, has_diagonal = FALSE)


# example 2: an upper triangle positive-definite matrix with 1s in diagonal
mat2 <- c(1.00,  0.32,  0.34, -0.09,  0.21,  0.30, 
          1.00,  0.45, -0.54,  0.54,  0.64,
          1.00, -0.26,  0.32,  0.34, 
          1.00, -0.52, -0.49, 
          1.00,  0.50,
          1.00)

check_if_positive_definite(mat2, is_lower_triangle = FALSE, has_diagonal = TRUE)


# the same matrix as the above, but one correlation changed so that the matrix is no longer positive-definite.
mat3 <- c(1.00,  0.32,  0.34, -0.09,  0.21, -0.90, # this top right corner correlation was change drom .30 to -.90
          1.00,  0.45, -0.54,  0.54,  0.64,
          1.00, -0.26,  0.32,  0.34, 
          1.00, -0.52, -0.49, 
          1.00,  0.50,
          1.00)

check_if_positive_definite(mat3, is_lower_triangle = FALSE, has_diagonal = TRUE)

}
