% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_consistency.R
\name{correlation_consistency}
\alias{correlation_consistency}
\title{correlation_consistency}
\usage{
correlation_consistency(r_XY, r_XZ, r_YZ = NA)
}
\arguments{
\item{r_XY}{The reported Pearson's r correlation variables X and Y.}

\item{r_XZ}{The reported Pearson's r correlation variables X and Z.}

\item{r_YZ}{The reported Pearson's r correlation variables Y and Z. This argument is optional and defaults to NA.}
}
\value{
a data frame with the variables the inputs (r_XY, r_XZ, and r_YZ), r_YZ_lower_bound and r_YZ_upper_bound (calculated from r_XY and r_XZ), and consistent (whether r_YZ is within the lower and upper bounds).
}
\description{
Examine the inequality of correlation coefficients. That is, given the correlations between X and Y (r_XY) and X and Z (r_XZ) in the same sample, determine (a) the mathematical bounds of the correlation between Y and Z (r_YX), and (b) whether the reported empirical correlation falls within these bounds.
}
\examples{
correlation_consistency(r_XY = .70, r_XZ = .80, r_YZ = .10)

correlation_consistency(r_XY = .90, r_XZ = .90, r_YZ = .10)

correlation_consistency(r_XY = .50, r_XZ = .40, r_YZ = .20)

}
