% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEM.R
\name{MEM}
\alias{MEM}
\title{Regression in model with errors}
\usage{
MEM(
  x,
  y,
  lambda_x,
  Correction = FALSE,
  method = "Bartlett",
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{Values vector}

\item{y}{Values vector}

\item{lambda_x}{Ratio of variances between values and error}

\item{Correction}{Small sample correction, use TRUE or FALSE}

\item{method}{The correction method that will be considered "Bartlett", "B1" or "B2"}

\item{conf.level}{Significance level for the hypothesis test}

\item{...}{Not always use all arguments}
}
\value{
Estimated Parameters \code{x} e \code{y}.
}
\description{
MEM is used to adjust models with errors in the variables.    It can be used to perform regressions on variables that have measurement errors.
}
\details{
Only use the conf.level, beta_til e method arguments when running equals "Y", otherwise these arguments will not affect the results.
}
\examples{
x = seq(1,100,length.out = 100)
y = sort(runif(100,1,10))

MEM(x, y, lambda_x = 2 )
MEM(x, y, lambda_x = 1, conf.level = 0.95 ,Correction = TRUE)

}
\references{
Bartlett, M.S. (1937), Properties of sufficiency and statistical tests,Proceedings of RoyalSociety of London A, 160, 268-282.

Fuller, S. (1987),Measurement Error Models. Wiley, New York.

Melo, T.F.N., Vasconcellos, K.L.P., Lemonte, A.J. (2009). Some restriction tests in a newclass of regression models for proportions.Computational Statistics and Data Analysis, 53,3972–3979.
}
\seealso{
\code{\link[stats]{lm}}
}
\author{
Kaíque Ferreira Henrique de Souza

Tatiane Ferreira do N. M. da Silva
}
