% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEV.R
\name{MEV}
\alias{MEV}
\title{Regression in model with errors}
\usage{
MEV(x, y, lambda_x, Correction = FALSE, method = "Bartlett",
  beta_til = 0, conf.level = 0.95, ...)
}
\arguments{
\item{x}{Values vector}

\item{y}{Error vector}

\item{lambda_x}{Ratio of variances between values and error}

\item{Correction}{Small sample correction, use TRUE or FALSE}

\item{method}{The correction method that will be considered "Bartlett", "B1" or "B2"}

\item{beta_til}{Parameter value to be tested}

\item{conf.level}{Significance level for the hypothesis test}

\item{...}{Not always use all arguments}
}
\value{
Estimated Parameters \code{x} e \code{y}.
}
\description{
MEVV is used to adjust models with errors in the variables.    It can be used to perform regressions on variables that have measurement errors.
}
\details{
Only use the conf.level, beta_til e method arguments when running equals "Y", otherwise these arguments will not affect the results.
}
\examples{
x = seq(1,100,length.out = 100)
y = sort(runif(100,1,10))

MEV(x, y, lambda_x = 2 )
MEV(x, y, lambda_x = 1, beta_til = 4 ,conf.level = 0.95 ,Correction = TRUE)

}
\references{
Aoki, R., Bolfarine, H. e Singer, J.M. (2001). Null intercept measurement error regressionmodels.TestA, 10, 441-457.

Aubin, E.C.Q. e Cordeiro, G.M. (2000). Bartlett-corrected tests for normal linear modelswhen the error covariance matrix is nonscalar.Communications in Statistics,  Theory andMethods,29, 2405-2426.

Barndorff–Nielsen, O.E. (1986), Inference on full or partial parameters, based on the stan-dardized signed log likelihood ratio,Biometrika, 73, 307-322.

Bartlett, M.S. (1937), Properties of sufficiency and statistical tests,Proceedings of RoyalSociety of London A, 160, 268-282.

Bayer, F.M., Cribari–Neto, F. (2013). Bartlett corrections in beta regression models.Journalof Statistical Planning and Inference, 143, 531–547.

Botter, D.A. e Cordeiro G.M. (1997). Bartlett corrections for generalized linear models withdispersion covariates.Communications in Statistics, Theory and Methods, 26, 279-307.

Chan,  L.K.&Mak,  T.K.,  On  the  maximum  likelihood  estimation  of  a  linear  structuralrelashionship  when  the  intercept  is  known,Journal  of  Multivariate  Analysis,  9,  304-313(1979).

Fuller, S. (1987),Measurement Error Models. Wiley, New York.

Melo, T.F.N., Vasconcellos, K.L.P., Lemonte, A.J. (2009). Some restriction tests in a newclass of regression models for proportions.Computational Statistics and Data Analysis, 53,3972–3979.
}
\seealso{
\code{\link[stats]{lm}}
}
\author{
Kaíque Ferreira Henrique de Souza

Tatiane Ferreira do N. M. da Silva
}
