% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_from_correlation_matrix.R
\name{regression_from_correlation_matrix}
\alias{regression_from_correlation_matrix}
\title{regression_from_correlation_matrix}
\usage{
regression_from_correlation_matrix(model, cor, n)
}
\arguments{
\item{model}{lavaan/Wilkinson notation for the regression model, e.g., Y ~ X1 + X2}

\item{cor}{A correlation matrix. See examples for how to convert the upper or lower triangle of a correlation matrix reported in an article into the full matrix.}

\item{n}{An integer specifying the sample size.}
}
\value{
a data frame containing standardized regression coefficients and their confidence intervals, and p values.
}
\description{
Fit a univeriable or multivariable regression using a covariance matrix as an input rather than the data itself. Note that only continuous variables can be used.
}
\examples{
# dependencies 
library(dplyr)
library(janitor)
library(lavaan)
library(knitr)
library(kableExtra)

## results extracted from Haller et al. (2022) 'To Help or Not to Help? Prosocial Behavior, Its Association With Well-Being, and Predictors of Prosocial Behavior During the Coronavirus Disease Pandemic'
n <- 9496

name <- c("Prosocial behavior", "Well-being", "Social Support", "Perceived Stress", "Psychological Flexibility", "Positive Affect")

name <- janitor::make_clean_names(name)

correlation <- 
  lavaan::getCov(x = c(
    1.00,  0.32,  0.34, -0.09,  0.21,  0.30, # Prosocial behavior
           1.00,  0.45, -0.54,  0.54,  0.64, # Well-being
                  1.00, -0.26,  0.32,  0.34, # Social Support
                         1.00, -0.52, -0.49, # Perceived Stress
                                1.00,  0.50, # Psychological Flexibility
                                       1.00  # Positive Affect
  ),
  lower = FALSE,
  diagonal = TRUE, 
  names = name
  )


extracted_results <-
  list(
    n = n,
    name = name,
    correlation = correlation
  )

# fit regression using correlation matrix
fit <- regression_from_correlation_matrix(
  model = 'prosocial_behavior ~ perceived_stress + positive_affect + psychological_flexibility + social_support',
  cor = extracted_results$correlation,
  n = extracted_results$n
)

# print results
fit |>
  # round variables for printing
  mutate(across(c(where(is.numeric), -p), janitor::round_half_up, digits = 2)) |>
  # print html table
  kable() |>
  kable_classic(full_width = FALSE)


}
